Make user validation and creation of accounts more explicit
Creating new accounts and validation of the user occurred in a side effect of
the selcted_identity field of a serializer. These calls need to be made, but
should be more explicit. A selected_idenity is a relic from a time before.

In addition, create_new_accounts shouldn't check whether it should actually do
its job, by calling user.is_valid first.