Add test for instance delete
This test isn't doing too much, because the actual deleting occurs in a celery
task. We are already using this task elsewhere. The method wrapped by the task
should be tested separately

I'm overriding the 'Accept' header to 'application/json'. Because without it,
rest framework renders the dashboard view. Since rest framework wants the
browser to render something for delete it replaces our 204 with a 200. This is
all unnecessary, we should be globally passing this accept header, because
right now we're not testing actual api responses as much as we're testing what
rest_framework stuffs in the data attribute of the response.